+++ HISTORY.md
@@ -2,16 +2,20 @@
 
 ## in development:
 * **New Features:**
-    * CSV-parsing auto-detects the most likely delimiter among comma, semi-colon, pipe and tab
 * **Notable Changes:**
 * **Notable Fixes:**
 * **Translation Updates:**
 
+## v5.0.3
+* **Notable Change:**
+    * CSV-parsing auto-detects the most likely delimiter among comma, semi-colon, pipe and tab
+
 ### 2019-07-10
 * threads: enhanced CSV-parsing to auto-detect the most likely delimiter among comma, semi-colon, pipe and tab
+* maintenance release
 
 ## v5.0.2
-* **New Features:**
+* **New Feature:**
     * experimental tolerant color collision detection (in dev mode)
 * **Notable Fixes:**
     * fixed a layout glitch for variadic C-shaped input slots, thanks, Bernat, for reporting it!